+++ lib/isox.dart
@@ -5,3 +5,4 @@ export 'package:isox/src/typedefs.dart';
 export 'package:isox/src/command.dart';
 export 'package:isox/src/config.dart';
 export 'package:isox/src/instance.dart';
+export 'package:isox/src/exceptions.dart';