+++ lib/src/config.dart
@@ -1,5 +1,4 @@
 import 'package:isox/isox.dart';
-import 'package:isox/src/config_implementation.dart';
 
 /// Describes the configuration for an Isox isolate. Using this config, the
 /// commands can be registered and an error handler can be set.
@@ -9,7 +8,7 @@ import 'package:isox/src/config_implementation.dart';
 abstract class IsoxConfig {
   /// Will add the given [command] to this config. The [command] must not be
   /// null. If there is already a command with the same name, a
-  ///[IsoxRegistryDuplicationException] will be thrown.
+  ///[IsoxCommandDuplicationException] will be thrown.
   void command(IsoxCommand<dynamic, dynamic, dynamic> command);
 
   /// Will set the [IsoxErrorHandler] for this config.